+++ src/main/java/roomescape/controller/member/MemberController.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.controller.member;
 
 import java.util.List;
 import org.springframework.http.HttpStatus;
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.dto.member.MemberResponseDto;
+import roomescape.controller.member.dto.MemberResponseDto;
 import roomescape.service.MemberService;
 
 @RestController
