+++ src/main/java/roomescape/service/AuthService.java
@@ -5,9 +5,8 @@ import org.springframework.stereotype.Service;
 import roomescape.common.exception.NotFoundException;
 import roomescape.common.exception.UnauthorizedException;
 import roomescape.dao.MemberDao;
-import roomescape.dto.auth.MemberInfoDto;
-import roomescape.dto.auth.LoginRequestDto;
-import roomescape.model.Member;
+import roomescape.controller.auth.dto.MemberInfoDto;
+import roomescape.controller.auth.dto.LoginRequestDto;
 import roomescape.common.util.JwtProvider;
 
 @Service
