+++ src/main/java/roomescape/service/ReservationService.java
@@ -9,12 +9,12 @@ import roomescape.dao.MemberDao;
 import roomescape.dao.ReservationDao;
 import roomescape.dao.ReservationTimeDao;
 import roomescape.dao.ThemeDao;
-import roomescape.dto.auth.MemberInfoDto;
-import roomescape.dto.reservation.AdminReservationRequestDto;
-import roomescape.dto.reservation.MemberReservationRequestDto;
-import roomescape.dto.reservation.ReservationResponseDto;
-import roomescape.dto.reservationtime.ReservationTimeResponseDto;
-import roomescape.dto.theme.ThemeResponseDto;
+import roomescape.controller.auth.dto.MemberInfoDto;
+import roomescape.controller.reservation.dto.AdminReservationRequestDto;
+import roomescape.controller.reservation.dto.MemberReservationRequestDto;
+import roomescape.controller.reservation.dto.ReservationResponseDto;
+import roomescape.controller.reservationTime.dto.ReservationTimeResponseDto;
+import roomescape.controller.theme.dto.ThemeResponseDto;
 import roomescape.model.Member;
 import roomescape.model.Reservation;
 import roomescape.model.ReservationTime;
