+++ src/test/java/roomescape/service/ReservationServiceTest.java
@@ -23,9 +23,9 @@ import roomescape.dao.MemberDao;
 import roomescape.dao.ReservationDao;
 import roomescape.dao.ReservationTimeDao;
 import roomescape.dao.ThemeDao;
-import roomescape.dto.auth.MemberInfoDto;
-import roomescape.dto.reservation.MemberReservationRequestDto;
-import roomescape.dto.reservation.ReservationResponseDto;
+import roomescape.controller.auth.dto.MemberInfoDto;
+import roomescape.controller.reservation.dto.MemberReservationRequestDto;
+import roomescape.controller.reservation.dto.ReservationResponseDto;
 import roomescape.model.Member;
 import roomescape.model.Reservation;
 import roomescape.model.ReservationTime;
