+++ src/test/java/roomescape/service/ThemeServiceTest.java
@@ -16,8 +16,8 @@ import roomescape.common.exception.DuplicatedException;
 import roomescape.common.exception.ResourceInUseException;
 import roomescape.dao.ReservationDao;
 import roomescape.dao.ThemeDao;
-import roomescape.dto.theme.ThemeRequestDto;
-import roomescape.dto.theme.ThemeResponseDto;
+import roomescape.controller.theme.dto.ThemeRequestDto;
+import roomescape.controller.theme.dto.ThemeResponseDto;
 import roomescape.model.Theme;
 
 @ExtendWith(MockitoExtension.class)
